+++ src/components/Team/index.tsx
@@ -1,41 +1,33 @@
-import { button, CallToAction, container } from 'components/CallToAction'
+import { CallToAction } from 'components/CallToAction'
 import { PineappleText } from 'components/Job/Sidebar'
-import Link from 'components/Link'
 import { InProgress } from 'components/Roadmap/InProgress'
 import { Question } from 'components/Squeak'
 import useTeamUpdates from 'hooks/useTeamUpdates'
 import { graphql, navigate, useStaticQuery } from 'gatsby'
 import { kebabCase } from 'lib/utils'
 import React, { useState } from 'react'
-import { UnderConsideration } from 'components/Roadmap/UnderConsideration'
-import { Change } from '../../templates/Changelog'
 import { MDXProvider } from '@mdx-js/react'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
 import { SmoothScroll } from 'components/Products/SmoothScroll'
 import Tooltip from 'components/Tooltip'
 import SEO from 'components/seo'
 import SideModal from 'components/Modal/SideModal'
-import { Avatar } from 'components/MainNav'
-import getAvatarURL from 'components/Squeak/util/getAvatar'
-import Markdown from 'markdown-to-jsx'
 import { AddTeamMember } from 'components/TeamMembers'
 import useTeam from 'hooks/useTeam'
-import { IconArrowLeft, IconInfo, IconSpinner, IconX } from '@posthog/icons'
+import { IconInfo, IconSpinner, IconX } from '@posthog/icons'
 import { useUser } from 'hooks/useUser'
 import { useFormik } from 'formik'
 import TeamUpdate from 'components/TeamUpdate'
 import usePostHog from '../../hooks/usePostHog'
 import { PrivateLink } from 'components/PrivateLink'
 import Stickers from 'components/ProfileStickers'
-import TeamPatch from 'components/TeamPatch'
-import CloudinaryImage from 'components/CloudinaryImage'
-import AutosizeInput from 'react-input-autosize'
-import ImageDrop from 'components/ImageDrop'
 import uploadImage from 'components/Squeak/util/uploadImage'
-import Modal from 'components/Modal'
-import Select from 'components/Select'
 import slugify from 'slugify'
 import * as yup from 'yup'
+import Section from './Section'
+import Header from './Header'
+import Profile, { ProfileData } from './Profile'
+import Roadmap from './Roadmap'
 
 const hedgehogImageWidth = 30
 const hedgehogLengthInches = 7
@@ -66,454 +58,19 @@ const SidebarSection = ({ title, tooltip, children }) => {
     )
 }
 
-const Section = ({ children, cta, title, className = '', id = '' }) => {
-    return (
-        <section id={id} className={`max-w-screen-xl mx-auto px-5 mt-6 mb-12 ${className}`}>
-            {title && (
-                <div className="flex flex-col md:flex-row justify-between items-baseline w-full mb-6 md:mb-8 relative after:h-px after:bg-border dark:after:bg-border-dark after:absolute after:top-1/2 after:left-0 after:w-full">
-                    <h4 className="m-0 bg-light dark:bg-dark relative z-10 pr-2">{title}</h4>
-                    {cta && (
-                        <aside className="bg-light dark:bg-dark relative z-10 md:pl-2 leading-tight -top-1">
-                            {cta}
-                        </aside>
-                    )}
-                </div>
-            )}
-            <div>{children}</div>
-        </section>
-    )
+interface TeamProps {
+    body: string
+    roadmaps?: any[]
+    objectives?: string
+    emojis?: any[]
+    newTeam?: boolean
+    slug: string
 }
 
-export const Profile = (profile) => {
-    const { firstName, lastName, country, companyRole, pineappleOnPizza, biography, isTeamLead, id, location, color } =
-        profile
-    const name = [firstName, lastName].filter(Boolean).join(' ')
-    return (
-        <div>
-            <div className="flex space-x-2 mb-6">
-                <Avatar
-                    className={`w-24 h-24 ${
-                        color ? `bg-${color}` : 'bg-accent dark:bg-dark'
-                    } rounded-full border border-border dark:border-dark`}
-                    src={getAvatarURL(profile)}
-                />
-                <div>
-                    <h2 className="m-0">{name}</h2>
-                    <p className="text-primary/50 text-sm dark:text-primary-dark/50 m-0">{companyRole}</p>
-                    <div className="flex space-x-1 items-center mt-1">
-                        <Stickers
-                            className="w-8 h-8"
-                            country={country}
-                            location={location}
-                            pineappleOnPizza={pineappleOnPizza}
-                            isTeamLead={isTeamLead}
-                        />
-                    </div>
-                </div>
-            </div>
-
-            {biography ? (
-                <Markdown className="bio-sidebar">{biography}</Markdown>
-            ) : (
-                <p className="bg-accent dark:bg-accent-dark border border-light dark:border-dark rounded p-4 text-sm">
-                    {firstName} has been too busy writing code to fill out a bio!
-                </p>
-            )}
-            <div className="mt-4">
-                <CallToAction
-                    to={`/community/profiles/${id}`}
-                    type="secondary"
-                    size="sm"
-                    width="full [&_span]:w-[calc(100%_+_3px)]"
-                >
-                    Visit full profile
-                </CallToAction>
-            </div>
-        </div>
-    )
-}
-
-const Description = ({ description, handleChange, values, editing }) => {
-    return editing ? (
-        <textarea
-            rows={5}
-            name="description"
-            onChange={handleChange}
-            placeholder="Description"
-            value={values.description}
-            className="w-full p-2 text-[15px] rounded-md bg-white dark:bg-accent-dark border border-border dark:border-dark mb-2 resize-none"
-        />
-    ) : description ? (
-        <p className="my-2 md:mb-4 text-[15px]" dangerouslySetInnerHTML={{ __html: description }} />
-    ) : null
-}
-
-const TeamName = ({ teamName, handleChange, values, editing }) => {
-    return (
-        <div className="mb-2">
-            {editing ? (
-                <div className="font-bold flex space-x-1 items-baseline">
-                    <AutosizeInput
-                        inputClassName="p-2 rounded-md bg-white dark:bg-accent-dark border border-border dark:border-dark"
-                        placeholder="Team name"
-                        type="text"
-                        name="name"
-                        onChange={handleChange}
-                        value={values.name}
-                    />
-                    <span>Team</span>
-                </div>
-            ) : (
-                <h1 className="m-0">{teamName} Team</h1>
-            )}
-        </div>
-    )
-}
-
-const TeamImage = ({ teamImage, editing, setFieldValue, values }) => {
-    const handleDrop = (image) => {
-        setFieldValue('teamImage', image)
-    }
-
-    return !!teamImage?.image?.data || editing ? (
-        <figure className="rotate-2 max-w-sm flex flex-col gap-2 mt-8 md:mt-0 ml-auto">
-            <div className="bg-accent aspect-video flex justify-center items-center shadow-xl border-8 border-white rounded-md">
-                {editing ? (
-                    <div className="w-96">
-                        <ImageDrop
-                            onRemove={() => setFieldValue('teamImage', null)}
-                            onDrop={handleDrop}
-                            image={values.teamImage}
-                        />
-                    </div>
-                ) : (
-                    <CloudinaryImage src={teamImage?.image?.data?.attributes?.url} className="" />
-                )}
-            </div>
-            {editing ? (
-                <input
-                    name="teamImageCaption"
-                    onChange={(e) => setFieldValue('teamImageCaption', e.target.value)}
-                    value={values.teamImageCaption}
-                    placeholder="Caption"
-                    className="p-2 text-[13px] rounded-md bg-white dark:bg-accent-dark border border-border dark:border-dark"
-                />
-            ) : (
-                <div className="text-right text-[13px] mr-2">{teamImage?.caption}</div>
-            )}
-        </figure>
-    ) : null
-}
-
-const CrestBuilderField = ({ label, name, value, setFieldValue, type = 'text' }) => {
-    return (
-        <div className="flex flex-col gap-1">
-            <label htmlFor={name} className="text-sm opacity-60 font-semibold block text-black dark:text-primary-dark">
-                {label}
-            </label>
-            <input
-                type={type}
-                id={name}
-                name={name}
-                value={value}
-                onChange={(e) => setFieldValue(name, e.target.value)}
-                className="border border-border dark:border-dark rounded-md p-2"
-                placeholder={label}
-            />
-        </div>
-    )
-}
-
-const Crest = ({ crest, crestOptions, teamName, editing, setFieldValue, values, updateTeam }) => {
-    const [crestBuilderOpen, setCrestBuilderOpen] = useState(false)
-    const { user, getJwt } = useUser()
-
-    const handleDrop = (image) => {
-        setFieldValue('crestImage', image)
-    }
-
-    return (
-        <>
-            <Modal open={crestBuilderOpen} setOpen={setCrestBuilderOpen}>
-                <div
-                    onClick={() => setCrestBuilderOpen(false)}
-                    className="flex flex-start justify-center items-center absolute w-full h-full p-4"
-                >
-                    <div
-                        className="max-w-2xl w-full bg-white dark:bg-dark rounded-md border border-border dark:border-dark relative p-4 grid grid-cols-2 gap-4"
-                        onClick={(e) => e.stopPropagation()}
-                    >
-                        <button
-                            onClick={() => setCrestBuilderOpen(false)}
-                            className="absolute top-0 right-0 translate-x-[33%] -translate-y-[33%] bg-white dark:bg-dark rounded-full border border-border dark:border-dark p-2"
-                        >
-                            <IconX className="size-4 opacity-60" />
-                        </button>
-                        <div className="flex items-center justify-center relative group">
-                            <TeamPatch
-                                name={teamName}
-                                imageUrl={values.crestImage?.objectURL}
-                                className="h-48 md:h-80 -mt-2 md:-mt-6 mb-2 md:mb-0"
-                                {...values.crestOptions}
-                            />
-                            {!values.crestImage?.objectURL ? (
-                                <div className="absolute">
-                                    <ImageDrop
-                                        className="!size-12 -ml-2"
-                                        onRemove={() => setFieldValue('crestImage', null)}
-                                        onDrop={handleDrop}
-                                        image={values.crestImage}
-                                    />
-                                </div>
-                            ) : (
-                                <button
-                                    onClick={() => setFieldValue('crestImage', null)}
-                                    className="group-hover:opacity-100 opacity-0 transition-opacity absolute size-8 bg-white rounded-full border border-border dark:border-dark flex items-center justify-center"
-                                >
-                                    <IconX className="size-4" />
-                                </button>
-                            )}
-                        </div>
-
-                        <div>
-                            <form className="grid gap-2">
-                                <div className="grid grid-cols-2 gap-2 gap-x-4">
-                                    <Select
-                                        className="!p-0"
-                                        options={[
-                                            { label: 'Black', value: 'black' },
-                                            { label: 'Brown', value: 'brown' },
-                                            { label: 'Navy', value: 'navy' },
-                                            { label: 'White', value: 'white' },
-                                        ]}
-                                        onChange={(value) => setFieldValue('crestOptions.textColor', value)}
-                                        value={values.crestOptions.textColor}
-                                        placeholder="Text color"
-                                    />
-                                    <Select
-                                        className="!p-0"
-                                        options={[
-                                            { label: 'Dark', value: 'dark' },
-                                            { label: 'Light', value: 'light' },
-                                        ]}
-                                        onChange={(value) => setFieldValue('crestOptions.textShadow', value)}
-                                        value={values.crestOptions.textShadow}
-                                        placeholder="Text shadow"
-                                    />
-                                    <Select
-                                        className="!p-0"
-                                        options={[
-                                            { label: 'Round', value: 'round' },
-                                            { label: 'Half Round', value: 'half-round' },
-                                            { label: 'Square', value: 'square' },
-                                            { label: 'Hexagon', value: 'hexagon' },
-                                            { label: 'Oval', value: 'oval' },
-                                            { label: 'Squareish', value: 'squareish' },
-                                            { label: 'Shield', value: 'shield' },
-                                        ]}
-                                        onChange={(value) => setFieldValue('crestOptions.frame', value)}
-                                        value={values.crestOptions.frame}
-                                        placeholder="Frame style"
-                                    />
-                                    <Select
-                                        className="!p-0"
-                                        options={[
-                                            { label: 'Blue', value: 'blue-2' },
-                                            { label: 'Burnt Orange', value: 'burnt-orange' },
-                                            { label: 'Creamsicle', value: 'creamsicle' },
-                                            { label: 'Fuchsia', value: 'fuchsia' },
-                                            { label: 'Green', value: 'green' },
-                                            { label: 'Gold', value: 'gold' },
-                                            { label: 'Light Blue', value: 'light-blue' },
-                                            { label: 'Light Purple', value: 'light-purple' },
-                                            { label: 'Light Yellow', value: 'light-yellow' },
-                                            { label: 'Navy', value: 'navy' },
-                                            { label: 'Orange', value: 'orange' },
-                                            { label: 'Pale Blue', value: 'pale-blue' },
-                                            { label: 'Pink', value: 'pink' },
-                                            { label: 'Purple 2', value: 'purple-2' },
-                                            { label: 'Red 2', value: 'red-2' },
-                                            { label: 'Teal 2', value: 'teal-2' },
-                                        ]}
-                                        onChange={(value) => setFieldValue('crestOptions.frameColor', value)}
-                                        value={values.crestOptions.frameColor}
-                                        placeholder="Frame color"
-                                    />
-                                    <Select
-                                        className="!p-0"
-                                        options={[
-                                            { label: 'Straight', value: 'straight' },
-                                            { label: 'Curved', value: 'curved' },
-                                            { label: 'Wavy', value: 'wavy' },
-                                            { label: 'Downward Curve', value: 'downward-curve' },
-                                            { label: 'Upward Curve', value: 'upward-curve' },
-                                            { label: 'Stepped', value: 'stepped' },
-                                        ]}
-                                        onChange={(value) => setFieldValue('crestOptions.plaque', value)}
-                                        value={values.crestOptions.plaque}
-                                        placeholder="Plaque style"
-                                    />
-                                    <Select
-                                        className="!p-0"
-                                        options={[
-                                            { label: 'Blue', value: 'blue-2' },
-                                            { label: 'Burnt Orange', value: 'burnt-orange' },
-                                            { label: 'Creamsicle', value: 'creamsicle' },
-                                            { label: 'Fuchsia', value: 'fuchsia' },
-                                            { label: 'Green', value: 'green' },
-                                            { label: 'Gold', value: 'gold' },
-                                            { label: 'Light Blue', value: 'light-blue' },
-                                            { label: 'Light Purple', value: 'light-purple' },
-                                            { label: 'Navy', value: 'navy' },
-                                            { label: 'Orange', value: 'orange' },
-                                            { label: 'Pale Blue', value: 'pale-blue' },
-                                            { label: 'Pink', value: 'pink' },
-                                            { label: 'Purple 2', value: 'purple-2' },
-                                            { label: 'Red 2', value: 'red-2' },
-                                            { label: 'Teal 2', value: 'teal-2' },
-                                            { label: 'White', value: 'white' },
-                                        ]}
-                                        onChange={(value) => setFieldValue('crestOptions.plaqueColor', value)}
-                                        value={values.crestOptions.plaqueColor}
-                                        placeholder="Plaque color"
-                                    />
-                                </div>
-                                <div className="flex flex-col gap-1">
-                                    <label className="text-sm opacity-60 font-semibold block text-black dark:text-primary-dark mb-1">
-                                        Image scale
-                                    </label>
-                                    <input
-                                        type="range"
-                                        min="50"
-                                        max="100"
-                                        step="5"
-                                        className="w-full accent-red dark:accent-yellow"
-                                        value={values.crestOptions.imageScale}
-                                        onChange={(e) => setFieldValue('crestOptions.imageScale', e.target.value)}
-                                    />
-                                </div>
-                                <div className="grid grid-cols-2 gap-2">
-                                    <div className="flex flex-col gap-1">
-                                        <label className="text-sm opacity-60 font-semibold block text-black dark:text-primary-dark mb-1">
-                                            Image X offset
-                                        </label>
-                                        <input
-                                            type="range"
-                                            min="-100"
-                                            max="100"
-                                            step="5"
-                                            className="w-full accent-red dark:accent-yellow"
-                                            value={values.crestOptions.imageXOffset}
-                                            onChange={(e) => setFieldValue('crestOptions.imageXOffset', e.target.value)}
-                                        />
-                                    </div>
-                                    <div className="flex flex-col gap-1">
-                                        <label className="text-sm opacity-60 font-semibold block text-black dark:text-primary-dark mb-1">
-                                            Image Y offset
-                                        </label>
-                                        <input
-                                            type="range"
-                                            min="-100"
-                                            max="100"
-                                            step="5"
-                                            className="w-full accent-red dark:accent-yellow"
-                                            value={values.crestOptions.imageYOffset}
-                                            onChange={(e) => setFieldValue('crestOptions.imageYOffset', e.target.value)}
-                                        />
-                                    </div>
-                                </div>
-                                <button
-                                    type="button"
-                                    onClick={() => setCrestBuilderOpen(false)}
-                                    className={`${container('primary', 'md', 'full')} mt-2`}
-                                >
-                                    <span className={button('primary', 'full', 'block', 'md')}>Save</span>
-                                </button>
-                            </form>
-                        </div>
-                    </div>
-                </div>
-            </Modal>
-
-            {values.crestOptions ? (
-                <div className="flex flex-col gap-2">
-                    <TeamPatch
-                        name={teamName}
-                        imageUrl={values.crestImage?.objectURL}
-                        className="h-48 md:h-80 -mt-2 md:-mt-6 mb-2 md:mb-0"
-                        {...values.crestOptions}
-                    />
-                    {editing && (
-                        <CallToAction onClick={() => setCrestBuilderOpen(true)} size="sm" type="secondary">
-                            Edit crest
-                        </CallToAction>
-                    )}
-                </div>
-            ) : editing ? (
-                <div className="size-60 border border-border dark:border-border-dark rounded-md flex items-center justify-center">
-                    <CallToAction onClick={() => setCrestBuilderOpen(true)} size="sm" type="secondary">
-                        Crest builder
-                    </CallToAction>
-                </div>
-            ) : null}
-        </>
-    )
-}
-
-const Header = ({
-    teamName,
-    crest,
-    crestOptions,
-    description,
-    teamImage,
-    hasInProgress,
-    editing,
-    handleChange,
-    values,
-    setFieldValue,
-    updateTeam,
-}) => {
-    return (
-        <Section className="mb-6">
-            <div className="flex flex-col md:flex-row space-x-4 items-center">
-                <Crest
-                    crest={crest}
-                    crestOptions={crestOptions}
-                    teamName={teamName}
-                    editing={editing}
-                    setFieldValue={setFieldValue}
-                    values={values}
-                    updateTeam={updateTeam}
-                />
-                <div className="max-w-xl w-full">
-                    <Link
-                        href="/teams"
-                        className="-ml-2 mb-1 inline-flex items-center gap-1 text-sm text-primary/50 dark:text-primary-dark/50 hover:text-primary dark:hover:text-primary-dark relative px-2 pt-1.5 pb-1 rounded hover:bg-light/50 hover:dark:bg-dark/50 border border-b-3 border-transparent md:hover:border-light dark:md:hover:border-dark hover:translate-y-[-1px] active:translate-y-[1px] active:transition-all"
-                    >
-                        <IconArrowLeft className="size-5" />
-                        <span>Teams</span>
-                    </Link>
-                    <TeamName teamName={teamName} handleChange={handleChange} values={values} editing={editing} />
-                    <Description
-                        description={description}
-                        handleChange={handleChange}
-                        values={values}
-                        editing={editing}
-                    />
-                    {hasInProgress && !editing && (
-                        <CallToAction type="secondary" size="md" to="#in-progress">
-                            See what we're building
-                        </CallToAction>
-                    )}
-                </div>
-                <TeamImage values={values} setFieldValue={setFieldValue} teamImage={teamImage} editing={editing} />
-            </div>
-        </Section>
-    )
-}
-
-export default function Team({ body, roadmaps, objectives, emojis, newTeam, slug }) {
+export default function Team({ body, roadmaps, objectives, emojis, newTeam, slug }: TeamProps): JSX.Element {
     const [saving, setSaving] = useState(false)
+    const [editing, setEditing] = useState(newTeam || false)
+    const [activeProfile, setActiveProfile] = useState<boolean | ProfileData>(false)
     const { team, updateTeam } = useTeam({
         slug,
     })
@@ -530,7 +87,7 @@ export default function Team({ body, roadmaps, objectives, emojis, newTeam, slug
             }
         }
     `)
-    const [editing, setEditing] = useState(newTeam)
+
     const { handleChange, values, submitForm, setFieldValue, errors, resetForm } = useFormik({
         enableReinitialize: true,
         validateOnMount: true,
@@ -614,7 +171,6 @@ export default function Team({ body, roadmaps, objectives, emojis, newTeam, slug
         },
     })
 
-    const teamName = name
     const teamLength = profiles?.data?.length
     const pineapplePercentage =
         teamLength &&
@@ -698,7 +254,6 @@ export default function Team({ body, roadmaps, objectives, emojis, newTeam, slug
     const hasUnderConsideration = underConsideration?.length > 0
     const hasInProgress = inProgress?.length > 0
     const hasBody = !!body
-    const [activeProfile, setActiveProfile] = useState(false)
     const heightToHedgehogs =
         profiles?.data?.reduce((acc, curr) => acc + (curr?.attributes?.height || 0), 0) / hedgehogLengthInches || 0
     const hedgehogPercentage =
@@ -720,21 +275,17 @@ export default function Team({ body, roadmaps, objectives, emojis, newTeam, slug
                 />
             )}
             <SideModal open={!!activeProfile} setOpen={setActiveProfile}>
-                <Profile {...activeProfile} />
+                {activeProfile && <Profile {...activeProfile} />}
             </SideModal>
             <Header
                 teamName={values.name}
-                crest={crest}
-                crestOptions={crestOptions}
                 description={description}
                 teamImage={teamImage}
                 hasInProgress={hasInProgress}
                 handleChange={handleChange}
                 values={values}
-                submitForm={submitForm}
                 editing={editing}
                 setFieldValue={setFieldValue}
-                updateTeam={updateTeam}
             />
             <SmoothScroll
                 menuItems={[
@@ -954,32 +505,11 @@ export default function Team({ body, roadmaps, objectives, emojis, newTeam, slug
                     </div>
                 </Section>
             )}
-            <div id="roadmap">
-                {hasUnderConsideration && (
-                    <Section title="Roadmap">
-                        <p className="-mt-2">
-                            Here’s what we’re considering building next. Vote for your favorites or share a new idea on{' '}
-                            <Link to="https://github.com/PostHog/posthog">GitHub</Link>.
-                        </p>
-                        <div className="max-w-2xl">
-                            <ul className="list-none m-0 p-0 space-y-4">
-                                {underConsideration.map((roadmap) => (
-                                    <UnderConsideration key={roadmap.squeakId} {...roadmap} />
-                                ))}
-                            </ul>
-                        </div>
-                    </Section>
-                )}
-                {recentlyShipped && (
-                    <Section title="Recently shipped">
-                        <div className="max-w-2xl team-page-content">
-                            <div className="border border-light dark:border-dark rounded bg-white dark:bg-accent-dark p-6">
-                                <Change {...recentlyShipped} />
-                            </div>
-                        </div>
-                    </Section>
-                )}
-            </div>
+            <Roadmap
+                hasUnderConsideration={hasUnderConsideration}
+                underConsideration={underConsideration}
+                recentlyShipped={recentlyShipped}
+            />
             {objectives && (
                 <Section title="Goals" id="goals">
                     <div className="article-content max-w-2xl team-page-content">
